Cleaned up the code and updated it.
